requirements.txt
lz4==4.0.2
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
pyparsing==3.0.9
pyzmq==23.2.0
